Extract XYZ, Lab and LCh color spaces
